--- cypress/integration/recieve-notice/newFile.js
@@ -1,1 +1,0 @@
-;
